requirements.txt
pydantic
peewee
langchain
sqlalchemy
pgvector
openai
tiktoken
python-telegram-bot
langdetect